Build react-select.js v1.0.0-rc.3
Summary:
This change simply builds `react-select.js` for changes in the head commit on master,
for use by webapp javascript.

This fixes warning when component is used in conjunction with React 16.3.0:
```
Warning: AutosizeInput: isMounted is deprecated. Instead, make sure to clean up subscriptions and pending requests in componentWillUnmount to prevent memory leaks.
```

I also tried upgrading to `v1.2.1` (latest), but this results in errors appearing in our
webapp javascript code that utilize this control. It's possible there may be
API changes that are incompatible? I didn't investigate this further yet, as `v1.0.0-rc.3`
seems to work for React 16.3.0 upgrade. However, it still has a dependency on `React.PropTypes`.

Test Plan:
 - Run `npm run build-ka`
 - Open `/react-sandbox` in webapp and verify no errors are logged in the browser console

Reviewers: emily, scottgrant

Reviewed By: emily

Subscribers: tom, john

Differential Revision: https://phabricator.khanacademy.org/D45468