Build react-select with webpack
Summary: As part of WEB-212 we need Kake to be able to `analyze_for_dependencies` on `react-select`. The current bundle we import is built with Browserify, which is full of calls to `require()` that cause Kake to choke. We'll just build this library with Webpack instead, which eliminates this problem -- as a nice side effect, we shave some size off of the bundle -- a preview of things to come!

Test Plan: Check a few callsites of `react-select` with the new bundle enabled, such as the React sandbox or a BigBingo experiment page.

Reviewers: #webpack, kevinb

Reviewed By: #webpack, kevinb

Subscribers: briangenisio, kevinb

Differential Revision: https://phabricator.khanacademy.org/D48778