Remove `browserify-shim` from `package.json`
Summary:
It's no longer needed.

We also are no longer minifying `react-select.js` in the root of this
project.  It is unnecessary as we minify our entire project before we
deploy.  Not minifying it means that it is easier to read and update in
the future should we need to.

This commit is part of the technical sustainability project migrating us
to use webpack instead of kake for bundling.

ADR: 101
Jira: WEB-222

Test Plan:
Run `npm run build-ka`
Open `/react-sandbox` in webapp and verify no errors are logged in the browser console

Reviewers: #webpack, kevinb, idr

Reviewed By: #webpack, kevinb, idr

Subscribers: miguel, kevinb

Differential Revision: https://phabricator.khanacademy.org/D48977